Move to `TTTReborn.Map` namespace
