+++ code/rounds/InProgressRound.cs
@@ -6,7 +6,7 @@ using Sandbox;
 
 using TTTReborn.Globals;
 using TTTReborn.Items;
-using TTTReborn.MapEntities;
+using TTTReborn.Map;
 using TTTReborn.Player;
 using TTTReborn.Roles;
 using TTTReborn.Teams;
